kiv: Add a button to switch scaling modes.
Two scaling modes (fit to the window, no scaling) were implemented
some time ago. One can switch them with not obvious shortcuts 'w' and
'=' (like in sxiv). This commit adds a button to the toolbar to do the
same.

Also, the window title now mentions the number of supported images in
the current directory, not the total number of files in it.

Align code to the kernel style guide.


git-svn-id: svn://kolibrios.org@7860 a494cfbc-eb01-0410-851d-a64ba20cac60